Create hashscript
Script that bring the user to a new page without changing the actual page. the user click a link we retrieve the link information and replace the page content. Some one would use this to keep a history of where the user is going, if the click back they can go back to all the elements of that page that were there before.